@@ -1,6 +1,6 @@
 {
   "name": "with-redux-wrapper",
-  "version": "2.0.0",
+  "version": "1.0.0",
   "scripts": {
     "dev": "next",
     "build": "next build",
@@ -9,12 +9,12 @@
   "dependencies": {
     "next": "latest",
     "next-redux-wrapper": "latest",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-redux": "5.0.7",
-    "redux": "4.0.0",
-    "redux-devtools-extension": "2.13.2",
-    "redux-thunk": "2.2.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-redux": "7.1.3",
+    "redux": "4.0.5",
+    "redux-devtools-extension": "2.13.8",
+    "redux-thunk": "2.3.0"
   },
   "license": "ISC"
 }
