@@ -1,33 +1,27 @@
-/* eslint-disable */
-import React, { Component } from 'react'
+import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
-import { addCount } from '../store'
+import { addCount } from '../store/count/action'
 
-class AddCount extends Component {
-  add = () => {
-    this.props.addCount()
-  }
-
-  render() {
-    const { count } = this.props
-    return (
-      <div>
-        <style jsx>{`
-          div {
-            padding: 0 0 20px 0;
-          }
-        `}</style>
-        <h1>
-          AddCount: <span>{count}</span>
-        </h1>
-        <button onClick={this.add}>Add To Count</button>
-      </div>
-    )
-  }
+const AddCount = ({ count, addCount }) => {
+  return (
+    <div>
+      <style jsx>{`
+        div {
+          padding: 0 0 20px 0;
+        }
+      `}</style>
+      <h1>
+        AddCount: <span>{count}</span>
+      </h1>
+      <button onClick={addCount}>Add To Count</button>
+    </div>
+  )
 }
 
-const mapStateToProps = ({ count }) => ({ count })
+const mapStateToProps = state => ({
+  count: state.count.count,
+})
 
 const mapDispatchToProps = dispatch => {
   return {
