@@ -1,26 +1,15 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import App from 'next/app'
 import withRedux from 'next-redux-wrapper'
-import { initStore } from '../store'
+import { initStore } from '../store/store'
 
-export default withRedux(initStore)(
-  class MyApp extends App {
-    static async getInitialProps({ Component, ctx }) {
-      return {
-        pageProps: Component.getInitialProps
-          ? await Component.getInitialProps(ctx)
-          : {},
-      }
-    }
+const MyApp = props => {
+  const { Component, pageProps, store } = props
+  return (
+    <Provider store={store}>
+      <Component {...pageProps} />
+    </Provider>
+  )
+}
 
-    render() {
-      const { Component, pageProps, store } = this.props
-      return (
-        <Provider store={store}>
-          <Component {...pageProps} />
-        </Provider>
-      )
-    }
-  }
-)
+export default withRedux(initStore)(MyApp)
